requirements.txt
Appium-Python-Client>=1.0.2
numpy>=1.19.2
Pillow>=7.2.0
pytest>=6.0.2
pytest-bdd>=4.0.1
requests>=2.24.0
selenium>=3.141.0
opencv-python>=4.4.0.42
pytesseract>=0.3.6
python-jenkins>=1.7.0
setuptools>=50.3.0
pyparsing>=2.4.7
urllib3>=1.25.10